requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=6.0.0,<7.0.0
pulumi-gcp>=7.0.0,<8.0.0
google-cloud-vision>=3.0.0
pillow>=10.0.0
boto3>=1.28.0